Moved find_or_create_customer_and_conversation to Conversations context
