+++ lib/chat_api/conversations.ex
@@ -8,6 +8,7 @@ defmodule ChatApi.Conversations do
 
   alias ChatApi.Accounts.Account
   alias ChatApi.Conversations.Conversation
+  alias ChatApi.Customers
   alias ChatApi.Customers.Customer
   alias ChatApi.Messages.Message
   alias ChatApi.Tags.{Tag, ConversationTag}
@@ -189,7 +190,7 @@ defmodule ChatApi.Conversations do
     |> Repo.all()
   end
 
-  @spec find_latest_conversation(binary(), map()) :: Conversation.t()
+  @spec find_latest_conversation(binary(), map()) :: Conversation.t() | nil
   def find_latest_conversation(account_id, filters) do
     Conversation
     |> where(^filter_where(filters))
@@ -506,6 +507,59 @@ defmodule ChatApi.Conversations do
     })
     |> Repo.update()
   end
+
+  @spec find_or_create_customer_and_conversation(String.t(), String.t()) ::
+          {:ok, Customers.Customer.t(), Conversation.t()}
+          | {:error, Ecto.Changeset.t()}
+  def find_or_create_customer_and_conversation(account_id, phone) do
+    now = DateTime.utc_now()
+    sms_source = "sms"
+
+    case Customers.list_customers(account_id, %{phone: phone}) do
+      [] ->
+        with {:ok, customer} <-
+               Customers.create_customer(%{
+                 phone: phone,
+                 account_id: account_id,
+                 first_seen: now,
+                 last_seen: now
+               }),
+             {:ok, conversation} <-
+               create_conversation(%{
+                 account_id: account_id,
+                 customer_id: customer.id,
+                 source: sms_source
+               }) do
+          {:ok, customer, conversation}
+        else
+          error ->
+            error
+        end
+
+      [customer] ->
+        with nil <-
+               find_latest_conversation(account_id, %{
+                 customer_id: customer.id,
+                 source: sms_source
+               }),
+             {:ok, conversation} <-
+               create_conversation(%{
+                 account_id: account_id,
+                 customer_id: customer.id,
+                 source: sms_source
+               }) do
+          {:ok, customer, conversation}
+        else
+          %Conversation{} = conversation ->
+            {:ok, customer, conversation}
+
+          error ->
+            error
+        end
+    end
+  end
+
+
 
   #####################
   # Private methods
